IVYPORTAL-17453 Improve performance when open case filter dialog - Fix exception when reset filter
