+++ AxonIvyPortal/portal/src/ch/ivy/addon/portalkit/dto/dashboard/taskcolumn/CategoryColumnModel.java
@@ -168,4 +168,8 @@ public class CategoryColumnModel extends TaskColumnModel {
   public boolean canQuickSearch() {
     return true;
   }
+  
+  public void initCategoryTree() {
+    this.userCategoryTree = TaskTreeUtils.buildRoot();
+  }
 }
