+++ AxonIvyPortal/portal/src/ch/ivy/addon/portalkit/util/TaskTreeUtils.java
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.primefaces.model.CheckboxTreeNode;
 import org.primefaces.model.TreeNode;
 
@@ -16,13 +15,12 @@ import com.axonivy.portal.components.service.IvyAdapterService;
 import ch.ivy.addon.portalkit.bo.CategoryNode;
 import ch.ivy.addon.portalkit.constant.PortalConstants;
 import ch.ivy.addon.portalkit.ivydata.searchcriteria.TaskCategorySearchCriteria;
-import ch.ivyteam.ivy.environment.Ivy;
 import ch.ivyteam.ivy.process.call.SubProcessCall;
 import ch.ivyteam.ivy.security.exec.Sudo;
 import ch.ivyteam.ivy.workflow.category.CategoryTree;
 import ch.ivyteam.ivy.workflow.query.TaskQuery;
 
-public class TaskTreeUtils {
+public class TaskTreeUtils extends TreeUtils{
 
   public static final String DELIMITER = "/";
 
@@ -92,34 +90,5 @@ public class TaskTreeUtils {
     nodeData.setValue(name);
     nodeData.setCategory(category);
     return nodeData;
-  }
-
-  private static CheckboxTreeNode<CategoryNode> buildRoot() {
-    CategoryNode nodeData = new CategoryNode();
-    nodeData.setValue(StringUtils.EMPTY);
-    nodeData.setCategory(StringUtils.EMPTY);
-    CheckboxTreeNode<CategoryNode> checkboxTreeNode = new CheckboxTreeNode<>(StringUtils.EMPTY, nodeData, null);
-    checkboxTreeNode.setExpanded(true);
-    checkboxTreeNode.setSelected(true);
-    buildAllCategoriesNode(checkboxTreeNode);
-    return checkboxTreeNode;
-  }
-
-  private static void buildAllCategoriesNode(CheckboxTreeNode<CategoryNode> root) {
-    CategoryNode nodeData = new CategoryNode();
-    nodeData.setValue(Ivy.cms().co("/ch.ivy.addon.portalkit.ui.jsf/common/allCategories"));
-    nodeData.setCategory(StringUtils.EMPTY);
-    CheckboxTreeNode<CategoryNode> checkboxTreeNode = new CheckboxTreeNode<>(nodeData, root);
-    checkboxTreeNode.setExpanded(true);
-    checkboxTreeNode.setSelected(true);
-    buildNoCategoryNode(checkboxTreeNode);
-  }
-
-  private static void buildNoCategoryNode(CheckboxTreeNode<CategoryNode> parent) {
-    CategoryNode nodeData = new CategoryNode();
-    nodeData.setValue(Ivy.cms().co(CategoryUtils.NO_CATEGORY_CMS));
-    nodeData.setCategory(CategoryUtils.NO_CATEGORY);
-    CheckboxTreeNode<CategoryNode> checkboxTreeNode = new CheckboxTreeNode<CategoryNode>(nodeData, parent);
-    checkboxTreeNode.setSelected(true);
-  }
+  }  
 }
